app/requirements.txt
pandas==2.2.1
psycopg2-binary==2.9.9
streamlit==1.32.0
streamlit_option_menu==0.3.12
